fixed all styling issues
