+++ src/App.css
@@ -1,5 +1,6 @@
 body {
     font-family: Asap;
     overflow-x: hidden;
-    background-color: #eef1f7
+    overflow-y: hidden;
+    background-color: #E5E5E5;
 }
