+++ src/components/Booking/Booking.css
@@ -49,7 +49,6 @@
     position: unset !important;
     text-align: center;
     color: #01194A;
-    font-weight: normal;
     font-size: 40px;
     line-height: 46px;
     font-style: normal;
