+++ src/components/Booking/Park/ParkingSpace/ParkSpace.css
@@ -11,7 +11,7 @@
 
 .available-slot {
     cursor: pointer;
-    background-color: #c4c4c4;
+    background-color: #C0C0C0;
 }
 
 .occupied-slot {
