+++ src/components/Booking/ParkDescription/ParkDescription.css
@@ -27,7 +27,7 @@
 }
 
 .spot-available {
-    background-color: #c4c4c4;
+    background-color: #C0C0C0;
 }
 
 .park-description h5 {
