+++ src/components/BookingHistory/BookingHistory.css
@@ -15,13 +15,13 @@
     align-items: center;
     flex-direction: column !important;
     top: 0;
-    bottom: 0;
     overflow-x: hidden;
     overflow-y: scroll;
     position: fixed;
     width: 90%;
     max-width: 90%;
     right: 0;
+    bottom: 30px;
 }
 
 
@@ -30,7 +30,6 @@
     text-align: center;
     padding-bottom: 40px;
     color: #01194A;
-    font-weight: normal;
     font-size: 40px;
     line-height: 46px;
     font-style: normal;
