+++ src/components/common/SideBar/SideBar.css
@@ -37,7 +37,7 @@
 }
 
 .sidebar-text-focus {
-    background-color: #eef1f7;
+    background-color: #E5E5E5;
     border-radius: 20px;
     color: #01194A;
 }
